rename: HomePart2.jsx를 HomePart.jsx로 수정
